LayoutTests/inspector/dom-debugger/url-breakpoints-matching-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for regex URL breakpoints.


== Running test suite: URLBreakpoint.BreakOnMatchingRegularExpression
-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.Fetch
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...
Request Fetch "resources/dataFetch.json"...
PAUSED:
PASS: Pause reason should be Fetch.
PASS: Pause data URL should match breakpoint URL.
CALL STACK:
0: [F] loadResourceFetch
1: [P] Global Code
-- Running test teardown.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.Fetch.Options.Condition
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.Fetch.Options.IgnoreCount
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.Fetch.Options.Action.Log
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Adding log action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.Fetch.Options.Actions.Evaluate
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Adding evaluate action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.XHR
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...
Request XHR "resources/dataXHR.json"...
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should match breakpoint URL.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.XHR.Options.Condition
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.XHR.Options.IgnoreCount
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.XHR.Options.Action.Log
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Adding log action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnMatchingRegularExpression.XHR.Options.Actions.Evaluate
Adding "regex:data[A-Z]*.(json|txt|png)" URL Breakpoint...

Adding evaluate action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.
